requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
python-multipart==0.0.26

langchain-openai==1.1.12
langchain-community==0.4.1
langchain-text-splitters==1.1.1

faiss-cpu==1.13.2
pypdf==6.10.1

# Testing (learning-friendly)
pytest==9.0.3

